backend/requirements.txt
# ── Core Framework ──
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ── AI / ML ──
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
numpy==1.26.2
scipy==1.11.4
pandas==2.1.4

# ── NLP ──
nltk==3.8.1

# ── PDF Processing ──
PyPDF2==3.0.1

# ── Image Processing ──
Pillow==10.1.0
torchvision==0.16.2

# ── Database ──
aiosqlite==0.19.0

# ── Utilities ──
pydantic==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
joblib==1.3.2
httpx==0.25.2

# ── PDF Report Generation ──
reportlab==4.0.8

# ── HTTP Client (Ollama) ──
httpx==0.25.2
